refactor: break apart user registration form from App.tsx
